requirements.txt
dotenv
azure-cognitiveservices-vision-customvision